Refactor the database manager class.
<h1>Rationale:</h1>

In preparation for adding experiences to the database, this commit significantly reduces the size of the manager by refactoring sections of code to other classes.  Some pruning has occurred as well.

<h1>File changes:</h1>

modified:   app/src/main/java/com/pajato/android/gamechat/account/Account.java

- getDisplayName(): simplify to lose a few lines of code.

modified:   app/src/main/java/com/pajato/android/gamechat/account/AccountManager.java

- getFirstName(): lose this, for now.
- onAuthStateChanged(): use the new database registrar to handle registering handlers.
- AccountChangeHandler#onDataChange(): simplify.
- AccountChangeHandler#getAccount(): build a partial account object.

modified:   app/src/main/java/com/pajato/android/gamechat/chat/BaseChatFragment.java

- joinDeveloperGroups(): use the new invitation manager.

modified:   app/src/main/java/com/pajato/android/gamechat/chat/ChatFragment.java

- onGroupProfileChange(): use the new invitation manager.

modified:   app/src/main/java/com/pajato/android/gamechat/chat/ChatListManager.java

- getGroupProfile(), onAccountStateChange(), onJoinedRoomsChange(), setMessageWatcher(): use the new database registrar to register handlers.

modified:   app/src/main/java/com/pajato/android/gamechat/database/DatabaseManager.java

- Summary: prune the code by refactoring and simplifying.
- Lose the obvious comments on the path and key constants.
- Move the handler registration pieces to the new registrar class.
- Move the invitation pieces to the new invitation manager.

modified:   app/src/main/java/com/pajato/android/gamechat/main/MainActivity.java

- init(): use the new registrar.

modified:   app/src/main/res/layout/fragment_ttt.xml
modified:   app/src/main/res/menu/chat_menu_base.xml
modified:   app/src/main/res/menu/drawer_body.xml
modified:   app/src/main/res/values/strings.xml

- Lose the @string/spaceValue strings.

new file:   app/src/main/java/com/pajato/android/gamechat/common/InvitationManager.java
new file:   app/src/main/java/com/pajato/android/gamechat/database/DatabaseRegistrar.java

- New files contained code factored out of the database manager.
