+++ app/src/main/java/com/pajato/android/gamechat/account/Account.java
@@ -76,16 +76,12 @@ import java.util.Map;
     // Public instance methods.
 
     /** Get a non-null display name using "Anonymous" if need be. */
-    public String getDisplayName(final Account current, final String me, final String anonymous) {
-        // Determine if the given account is this account and return the me value if it has been
-        // provided.
-        if (current != null && current.id.equals(id) && me != null) return me;
-
+    @Exclude public String getDisplayName(final String defaultName) {
         // Determine if this account has a display name or should use the given default.
-        if (displayName == null && anonymous != null) return anonymous;
+        if (displayName != null) return displayName;
 
-        // Otherwise use this display name.
-        return displayName;
+        // Otherwise use this default.
+        return defaultName;
     }
 
     /** Generate the map of data to persist into Firebase. */
